+++ src/main/java/com/example/healthgenie/boundedContext/community/comment/repository/CommentRepository.java
@@ -1,6 +1,6 @@
-package com.example.healthgenie.boundedContext.community.repository;
+package com.example.healthgenie.boundedContext.community.comment.repository;
 
-import com.example.healthgenie.boundedContext.community.entity.Comment;
+import com.example.healthgenie.boundedContext.community.comment.entity.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
