+++ src/main/java/com/example/healthgenie/boundedContext/community/post/entity/Post.java
@@ -1,6 +1,8 @@
-package com.example.healthgenie.boundedContext.community.entity;
+package com.example.healthgenie.boundedContext.community.post.entity;
 
 import com.example.healthgenie.base.entity.BaseEntity;
+import com.example.healthgenie.boundedContext.community.comment.entity.Comment;
+import com.example.healthgenie.boundedContext.community.photo.entity.Photo;
 import com.example.healthgenie.boundedContext.user.entity.User;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotBlank;
