+++ src/test/java/com/example/healthgenie/util/TestKrUtils.java
@@ -5,14 +5,14 @@ import com.example.healthgenie.boundedContext.chat.dto.ChatRoomRequest;
 import com.example.healthgenie.boundedContext.chat.entity.ChatRoom;
 import com.example.healthgenie.boundedContext.chat.entity.ChatRoomUser;
 import com.example.healthgenie.boundedContext.chat.repository.ChatRoomRepository;
-import com.example.healthgenie.boundedContext.community.dto.CommentRequest;
-import com.example.healthgenie.boundedContext.community.dto.PostRequest;
-import com.example.healthgenie.boundedContext.community.entity.Comment;
-import com.example.healthgenie.boundedContext.community.entity.Post;
-import com.example.healthgenie.boundedContext.community.entity.Photo;
-import com.example.healthgenie.boundedContext.community.repository.CommentRepository;
-import com.example.healthgenie.boundedContext.community.repository.PhotoRepository;
-import com.example.healthgenie.boundedContext.community.repository.PostRepository;
+import com.example.healthgenie.boundedContext.community.comment.dto.CommentRequest;
+import com.example.healthgenie.boundedContext.community.post.dto.PostRequest;
+import com.example.healthgenie.boundedContext.community.comment.entity.Comment;
+import com.example.healthgenie.boundedContext.community.post.entity.Post;
+import com.example.healthgenie.boundedContext.community.photo.entity.Photo;
+import com.example.healthgenie.boundedContext.community.comment.repository.CommentRepository;
+import com.example.healthgenie.boundedContext.community.photo.repository.PhotoRepository;
+import com.example.healthgenie.boundedContext.community.post.repository.PostRepository;
 import com.example.healthgenie.boundedContext.matching.dto.MatchingCondition;
 import com.example.healthgenie.boundedContext.matching.dto.MatchingRequest;
 import com.example.healthgenie.boundedContext.matching.dto.MatchingResponse;
